Pinionszek_API/Services/DatabaseServices/UserService/UserApiService.cs
using Microsoft.EntityFrameworkCore;
using Pinionszek_API.DbContexts;
using Pinionszek_API.Models.DatabaseModel;

namespace Pinionszek_API.Services.DatabaseServices.UserService
{
    public class UserApiService : IUserApiService
    {
        private readonly ProdDbContext _dbContext;

        public UserApiService(ProdDbContext dbContext)
        {
            _dbContext = dbContext;
        }

        public async Task<IEnumerable<DetailedCategory>?> GetUserCategoriesAsync(int idUser)
        {
            throw new NotImplementedException();
        }
    }
}
